extend access to actions

+++ src/Controller/AuctionController.php
@@ -7,10 +7,10 @@ use App\Form\AuctionType;
 use App\Form\BidType;
 use Symfony\Component\Form\Extension\Core\Type\SubmitType;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\Request;
-
 
 class AuctionController extends AbstractController
 {
@@ -90,7 +90,7 @@ class AuctionController extends AbstractController
     {
         $this->denyAccessUnlessGranted("ROLE_USER");
 
-        $auction = new Auction();
+        $auction = new Auction(); 
 
         $form = $this->createForm(AuctionType::class, $auction);
 
@@ -99,7 +99,8 @@ class AuctionController extends AbstractController
 
             if($form->isValid()){
                 $auction
-                ->setStatus(Auction::STATUS_ACTIVE);
+                ->setStatus(Auction::STATUS_ACTIVE)
+                ->setOwner($this->getUser());
 
                 $entityManager = $this->getDoctrine()->getManager();
                 $entityManager->persist($auction);
@@ -126,6 +127,12 @@ class AuctionController extends AbstractController
      */
     public function editAction(Request $request, Auction $auction)
     {
+        $this->denyAccessUnlessGranted("ROLE_USER");
+
+        if ($this->getUser() !== $auction->getOwner()){
+            throw new AccessDeniedException(); 
+        } 
+
         $form = $this->createForm(AuctionType::class, $auction);
 
         if ($request->isMethod("post")) {
@@ -157,6 +164,12 @@ class AuctionController extends AbstractController
      */
     public function deleteAction(Auction $auction)
     {
+        $this->denyAccessUnlessGranted("ROLE_USER");
+
+        if ($this->getUser() !== $auction->getOwner()){
+            throw new AccessDeniedException(); 
+        } 
+
         $entityManager = $this->getDoctrine()->getManager();
         $entityManager->remove($auction);
         $entityManager->flush();
@@ -175,6 +188,12 @@ class AuctionController extends AbstractController
      */
     public function finishAction(Auction $auction)
     {
+        $this->denyAccessUnlessGranted("ROLE_USER");
+
+        if ($this->getUser() !== $auction->getOwner()){
+            throw new AccessDeniedException(); 
+        } 
+
         $auction
             ->setExpiresAt(new \DateTime())
             ->setStatus(Auction::STATUS_FINISHED);
